test(comptime): update existing tests for new ValueType and add narrow type validation tests
Migrate existing comptime tests from .integer/.float to .i64/.f64 and
update expected error messages. Add 5 new comptime error tests:
u8 overflow, i8 overflow, unsigned negative, invalid u16, f32 overflow.

+++ test/comptime/default_hex_float.zig
@@ -6,7 +6,7 @@ comptime {
     const cmd = parsz.Command{
         .name = "test",
         .args = &.{
-            .{ .name = "ratio", .kind = .option, .value_type = .float, .long = "ratio", .default = "0x1.fp10" },
+            .{ .name = "ratio", .kind = .option, .value_type = .f64, .long = "ratio", .default = "0x1.fp10" },
         },
     };
     _ = parsz.ParseResult(cmd);
